% Auto-generated: do not edit by hand
\name{select}

\alias{select}

\title{Select component}

\description{

}

\usage{
select(id=NULL, class_name=NULL, disabled=NULL, multiple=NULL,
options=NULL, size=NULL, style=NULL, value=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{class_name}{Character. }

\item{disabled}{Logical. }

\item{multiple}{Logical. }

\item{options}{List of character | numerics. }

\item{size}{Numeric. }

\item{style}{Named list. }

\item{value}{Unnamed list | character | numeric. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

